+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
-## 4.7.0
+## 4.9.0
 - [#242](https://github.com/cohere-ai/cohere-python/pull/242)
   - Add token count to the chat response
+
+## 4.8.0
+- [#237](https://github.com/cohere-ai/cohere-python/pull/237)
+  - add support for custom models
+
+## 4.7.0
+- [#232](https://github.com/cohere-ai/cohere-python/pull/232)
+  - Support model parameter in tokenize and detokenize
 
 ## 4.6.1
 - [#240](https://github.com/cohere-ai/cohere-python/pull/240)